Enable E501 (line too long) rule and apply Black formatting
- Remove E501 from ignored rules in pyproject.toml
- Run Black formatter to fix most line length violations (486 -> 290)
- Black reformatted 33 files to comply with 100-character line limit
- Remaining violations will require manual intervention

+++ java-benchmark/src/main/java/com/fraiseql/benchmark/graphql/UserResolver.java
@@ -20,7 +20,7 @@ public class UserResolver {
     }
     
     @QueryMapping
-    public User user(@Argument Long id) {
+    public User user(@Argument Integer id) {
         return userService.findById(id);
     }
