+++ java-benchmark/src/main/java/com/fraiseql/benchmark/repository/PostRepository.java
@@ -9,10 +9,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface PostRepository extends JpaRepository<Post, Long> {
+public interface PostRepository extends JpaRepository<Post, Integer> {
     @Query("SELECT p FROM Post p JOIN FETCH p.author WHERE p.author.id = :userId")
-    List<Post> findByAuthorIdWithAuthor(@Param("userId") Long userId);
+    List<Post> findByAuthorIdWithAuthor(@Param("userId") Integer userId);
     
     @Query("SELECT p FROM Post p LEFT JOIN FETCH p.comments c LEFT JOIN FETCH c.author WHERE p.id IN :ids")
-    List<Post> findByIdsWithCommentsAndAuthors(@Param("ids") List<Long> ids);
+    List<Post> findByIdsWithCommentsAndAuthors(@Param("ids") List<Integer> ids);
 }
