+++ java-benchmark/src/main/java/com/fraiseql/benchmark/repository/UserRepository.java
@@ -10,12 +10,12 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
     
     @Query("SELECT DISTINCT u FROM User u LEFT JOIN FETCH u.posts WHERE u.id IN :ids")
-    List<User> findByIdsWithPosts(@Param("ids") List<Long> ids);
+    List<User> findByIdsWithPosts(@Param("ids") List<Integer> ids);
     
     @Query("SELECT u FROM User u LEFT JOIN FETCH u.posts p LEFT JOIN FETCH p.comments WHERE u.id = :id")
-    Optional<User> findByIdWithPostsAndComments(@Param("id") Long id);
+    Optional<User> findByIdWithPostsAndComments(@Param("id") Integer id);
 }
